Add files via upload
These are the tasks assigned to me as a Data Science Intern at Oasis Infobyte.

* Task 1 - IRIS FLOWER CLASSIFICATION
Iris flower has three species; setosa, versicolor, and virginica, which differs according to their
measurements. Now assume that you have the measurements of the iris flowers according to
their species, and here your task is to train a machine learning model that can learn from the
measurements of the iris species and classify them.

* Task 2 - UNEMPLOYMENT ANALYSIS WITH PYTHON
Unemployment is measured by the unemployment rate which is the number of people
who are unemployed as a percentage of the total labour force. We have seen a sharp
increase in the unemployment rate during Covid-19, so analyzing the unemployment rate
can be a good data science project. 

*Task 3 - CAR PRICE PREDICTION WITH MACHINE LEARNING
The price of a car depends on a lot of factors like the goodwill of the brand of the car,
features of the car, horsepower and the mileage it gives and many more. Car price
prediction is one of the major research areas in machine learning. So if you want to learn
how to train a car price prediction model then this project is for you.